@@ -4,6 +4,7 @@
   "version": "5.0.3",
   "description": "docker package",
   "main": "index.js",
+  "type": "module",
   "files": [
     "lib/",
     "README.md",
@@ -13,7 +14,7 @@
     "index.js"
   ],
   "scripts": {
-    "test": "docker-compose -f compose/base.yml -f compose/test.yml up --exit-code-from semantic-release --force-recreate --remove-orphans --build",
+    "test": "docker-compose -f compose/base.yml -f compose/test.yml up --exit-code-from semantic-release --remove-orphans --build",
     "start": "docker-compose -f compose/base.yml -f compose/dev.yml up --force-recreate --remove-orphans --build",
     "stop": "docker-compose -f compose/base.yml -f compose/dev.yml down",
     "tap": "tap",
@@ -37,15 +38,8 @@
     "url": "git+ssh://git@github.com/esatterwhite/semantic-release-docker.git"
   },
   "tap": {
-    "ts": false,
-    "jsx": false,
     "timeout": 60,
     "browser": false,
-    "check-coverage": true,
-    "lines": 95,
-    "branches": 95,
-    "statements": 95,
-    "functions": 95,
     "files": [
       "test/unit",
       "test/integration"
@@ -56,13 +50,6 @@
       "json",
       "json-summary",
       "html"
-    ],
-    "nyc-arg": [
-      "--exclude=coverage/",
-      "--exclude=test/",
-      "--exclude=.eslintrc.js",
-      "--exclude=release.config.js",
-      "--all"
     ]
   },
   "devDependencies": {
@@ -70,17 +57,19 @@
     "@semantic-release/changelog": "^5.0.1",
     "@semantic-release/git": "^9.0.0",
     "@semantic-release/github": "^7.0.7",
-    "eslint": "^8.5.0",
+    "eslint": "^9.13.0",
     "eslint-config-codedependant": "^3.0.0",
     "semantic-release": "17",
     "sinon": "^9.0.2",
-    "tap": "^16.0.0"
+    "tap": "^21.0.0"
   },
   "dependencies": {
     "@semantic-release/error": "^3.0.0",
     "debug": "^4.1.1",
-    "execa": "^4.0.2",
+    "execa": "^9.5.1",
+    "globals": "^15.11.0",
     "handlebars": "^4.7.7",
     "semver": "^7.3.2"
-  }
+  },
+  "packageManager": "pnpm@9.12.2+sha512.22721b3a11f81661ae1ec68ce1a7b879425a1ca5b991c975b074ac220b187ce56c708fe5db69f4c962c989452eee76c82877f4ee80f474cebd61ee13461b6228"
 }
