@@ -1,21 +1,21 @@
 'use strict'
 
-const crypto = require('crypto')
-const dockerPrepare = require('./lib/prepare.js')
-const dockerVerify = require('./lib/verify.js')
-const dockerPublish = require('./lib/publish.js')
-const buildConfig = require('./lib/build-config.js')
-const dockerSuccess = require('./lib/success.js')
-const dockerFail = require('./lib/fail.js')
+import crypto from 'crypto'
+import { dockerPrepare } from './lib/prepare.js'
+import { verify as dockerVerify } from './lib/verify.js'
+import { publish as dockerPublish } from './lib/publish.js'
+import { buildConfig } from './lib/build-config.js'
+import { success as dockerSuccess } from './lib/success.js'
+import { fail as dockerFail } from './lib/fail.js'
 const build_id = crypto.randomBytes(10).toString('hex')
 
-module.exports = {
+export {
   buildConfig
-, fail
-, prepare
-, publish
-, success
-, verifyConditions
+  , fail
+  , prepare
+  , publish
+  , success
+  , verifyConditions
 }
 
 async function fail(config, context) {
