@@ -1,12 +1,12 @@
 'use strict'
 
-const os = require('os')
-const path = require('path')
-const crypto = require('crypto')
-const execa = require('execa')
-const buildTemplateVars = require('../build-template-vars.js')
-const array = require('../lang/array/index.js')
-const string = require('../lang/string/index.js')
+import os from 'os'
+import path from 'path'
+import crypto from 'crypto'
+import { execa } from 'execa'
+import { buildTemplateVars } from '../build-template-vars.js'
+import * as array from '../lang/array/index.js'
+import * as string from '../lang/string/index.js'
 const SHA_REGEX = /(?:writing image\s)?[^@](?:sha\d{3}):(?<sha>\w+)/i
 
 function render(item, vars) {
@@ -18,23 +18,23 @@ function render(item, vars) {
   return string.template(item)(vars)
 }
 
-class Image {
+export class Image {
   constructor(opts) {
     const {
       registry = null
-    , project = null
-    , name = null
-    , sha = null
-    , build_id = crypto.randomBytes(10).toString('hex')
-    , dockerfile = 'Dockerfile'
-    , cwd = process.cwd()
-    , context = '.'
-    , network = 'default'
-    , publish = true
-    , quiet = true
-    , dry_run = false
-    , tags = []
-    , platform = []
+      , project = null
+      , name = null
+      , sha = null
+      , build_id = crypto.randomBytes(10).toString('hex')
+      , dockerfile = 'Dockerfile'
+      , cwd = process.cwd()
+      , context = '.'
+      , network = 'default'
+      , publish = true
+      , quiet = true
+      , dry_run = false
+      , tags = []
+      , platform = []
     } = opts || {}
 
     if (!name || typeof name !== 'string') {
@@ -45,19 +45,19 @@ class Image {
     this.sha = sha
     this.opts = {
       build_id: build_id
-    , args: new Map()
-    , context: context
-    , cwd: cwd
-    , dockerfile: dockerfile
-    , flags: new Map()
-    , name: name
-    , tags: tags
-    , network: network
-    , project: project
-    , registry: registry
-    , dry_run: dry_run
-    , publish: publish
-    , platform: array.toArray(platform)
+      , args: new Map()
+      , context: context
+      , cwd: cwd
+      , dockerfile: dockerfile
+      , flags: new Map()
+      , name: name
+      , tags: tags
+      , network: network
+      , project: project
+      , registry: registry
+      , dry_run: dry_run
+      , publish: publish
+      , platform: array.toArray(platform)
     }
 
     if (quiet) this.flag('quiet', null)
@@ -74,20 +74,20 @@ class Image {
       return string.template(template)(vars)
     }).filter(Boolean)
 
-    const image = new(this)({
+    const image = new (this)({
       registry: opts.registry
-    , project: opts.project
-    , name: opts.name
-    , dockerfile: opts.dockerfile
-    , build_id: opts.build
-    , cwd: context.cwd
-    , tags: tags
-    , context: opts.context
-    , network: opts.network
-    , quiet: opts.quiet
-    , publish: opts.publish
-    , platform: opts.platform
-    , dry_run: !!opts.dry_run
+      , project: opts.project
+      , name: opts.name
+      , dockerfile: opts.dockerfile
+      , build_id: opts.build
+      , cwd: context.cwd
+      , tags: tags
+      , context: opts.context
+      , network: opts.network
+      , quiet: opts.quiet
+      , publish: opts.publish
+      , platform: opts.platform
+      , dry_run: !!opts.dry_run
     })
     for (const [key, value] of Object.entries(opts.args)) {
       image.arg(key, string.template(value)(vars))
@@ -127,7 +127,7 @@ class Image {
 
   set context(ctx) {
     this.opts.context = ctx
-    return this.opts.context
+    // return this.opts.context
   }
 
   get dockerfile() {
@@ -182,18 +182,18 @@ class Image {
   get docker_cmd() {
     return [
       'build'
-    , `--network=${this.network}`
-    , '--tag'
-    , this.name
-    , ...this.flags
-    , '-f'
-    , this.dockerfile
-    , this.context
+      , `--network=${this.network}`
+      , '--tag'
+      , this.name
+      , ...this.flags
+      , '-f'
+      , this.dockerfile
+      , this.context
     ].filter(Boolean)
   }
 
   get buildx_cmd() {
-    if (!this.is_buildx) return
+    if (!this.is_buildx) return undefined
     this.opts.flags.delete('provenance') // incompatible with load/push
     this.opts.flags.delete('output') // alias of load/push
     this.opts.flags.delete('load')
@@ -218,7 +218,7 @@ class Image {
   }
 
   arg(key, val = null) {
-    if (val === true || val == null) { // eslint-disable-line no-eq-null
+    if (val === true || val == null) {
       this.flag('build-arg', key)
     } else {
       this.flag('build-arg', `${key}=${val}`)
@@ -249,15 +249,15 @@ class Image {
     if (this.is_buildx) return
     const stream = execa('docker', [
       'run'
-    , '--rm'
-    , this.name
-    , ...array.toArray(cmd)
-    ], {all: true})
+      , '--rm'
+      , this.name
+      , ...array.toArray(cmd)
+    ], { all: true })
 
     stream.stdout.pipe(process.stdout)
     stream.stderr.pipe(process.stderr)
 
-    const {all} = await stream
+    const { all } = await stream
     return all
   }
 
@@ -265,7 +265,7 @@ class Image {
     const stream = execa('docker', this.build_cmd)
     stream.stdout.pipe(process.stdout)
     stream.stderr.pipe(process.stderr)
-    const {stdout, stderr} = await stream
+    const { stdout, stderr } = await stream
     const lines = (stdout || stderr).split(os.EOL)
     const len = lines.length - 1
 
@@ -304,10 +304,8 @@ class Image {
   }
 
   async clean() {
-    const {stdout: images} = await execa('docker', ['images', this.repo, '-q'])
+    const { stdout: images } = await execa('docker', ['images', this.repo, '-q'])
     if (!images) return
     await execa('docker', ['rmi', '-f', ...images.split(os.EOL)])
   }
 }
-
-module.exports = Image
