@@ -1,9 +1,7 @@
 'use strict'
 
-const postPublish = require('./post-publish.js')
+import { postPublish } from './post-publish.js'
 
-module.exports = fail
-
-function fail(opts, context) {
+export function fail(opts, context) {
   return postPublish(opts, context)
 }
