@@ -1,7 +1,5 @@
 'use strict'
 
-module.exports = eq
-
-function eq(lh, rh) {
+export function eq(lh, rh) {
   return lh === rh
 }
