@@ -1,7 +1,5 @@
 'use strict'
 
-module.exports = gt
-
-function gt(lh, rh) {
+export function gt(lh, rh) {
   return lh > rh
 }
