@@ -1,7 +1,5 @@
 'use strict'
 
-module.exports = gte
-
-function gte(lh, rh) {
+export function gte(lh, rh) {
   return lh >= rh
 }
