@@ -1,8 +1,6 @@
 'use strict'
 
-module.exports = includes
-
-function includes(input, arg, position) {
+export function includes(input, arg, position) {
   if (!Array.isArray(input) && typeof input !== 'string') return false
   return input.includes(arg, position)
 }
