@@ -1,8 +1,6 @@
 'use strict'
 
-module.exports = lower
-
-function lower(str) {
+export function lower(str) {
   if (typeof str !== 'string') return ''
   return str.toLowerCase()
 }
