@@ -1,7 +1,5 @@
 'use strict'
 
-module.exports = lt
-
-function lt(lh, rh) {
+export function lt(lh, rh) {
   return lh < rh
 }
