@@ -1,7 +1,5 @@
 'use strict'
 
-module.exports = lte
-
-function lte(lh, rh) {
+export function lte(lh, rh) {
   return lh < rh
 }
