@@ -1,9 +1,5 @@
 'use strict'
 
-module.exports = neq
-
-function neq(lh, rh) {
-  /* eslint-disable eqeqeq */
+export function neq(lh, rh) {
   return lh != rh
-  /* eslint-enable eqeqeq */
 }
