@@ -1,8 +1,6 @@
 'use strict'
 
-module.exports = pick
-
-function pick(...args) {
+export function pick(...args) {
   return args.find((value) => {
     /* eslint-disable no-eq-null */
     if (value == null) return false
