@@ -1,9 +1,7 @@
 
 'use strict'
 
-module.exports = upper
-
-function upper(str) {
+export function upper(str) {
   if (typeof str !== 'string') return ''
   return str.toUpperCase()
 }
