@@ -1,10 +1,13 @@
 'use strict'
 
-const Handlebars = require('handlebars')
-const helpers = require('./helpers')
+import Handlebars from 'handlebars'
+import * as helpers from './helpers/index.js'
+
 const handlebars = Handlebars.noConflict()
 
-handlebars.registerHelper(helpers)
+for (const [key, value] of Object.entries(helpers)) {
+  handlebars.registerHelper(key, value)
+}
 
-module.exports = handlebars
+export { handlebars }
 
