@@ -1,8 +1,6 @@
 'use strict'
 
-module.exports = getProperty
-
-function getProperty(obj, str, sep = '.') {
+export function getProperty(obj, str, sep = '.') {
   if (!obj || !str) return null
   const parts = str.split(sep)
   let ret = obj
