@@ -1,8 +1,6 @@
 'use strict'
 
-module.exports = hasProperty
-
-function hasProperty(obj, prop) {
+export function hasProperty(obj, prop) {
   if (!obj) return false
   return Object.prototype.hasOwnProperty.call(obj, prop)
 }
