@@ -1,10 +1,8 @@
 'use strict'
 
-const handlebars = require('../../handlebars')
+import { handlebars } from '../../handlebars/index.js'
 
-module.exports = template
-
-function template(str) {
+export function template(str) {
   if (typeof str !== 'string') return echo(str)
   return handlebars.compile(str)
 }
