@@ -5,7 +5,7 @@
  * @author Eric Satterwhite
  **/
 
-module.exports = function typecast(value) {
+export function typeCast(value) {
   if (value === 'null' || value === null) return null
   if (value === 'undefined' || value === undefined) return undefined
   if (value === 'true' || value === true) return true
