@@ -2,11 +2,9 @@
 
 const NAME_EXP = /^(?:@([^/]+?)[/])?([^/]+?)$/
 
-module.exports = parsePkgName
-
-function parsePkgName(pkgname) {
-  if (!pkgname) return {scope: null, name: null}
+export function parsePkgName(pkgname) {
+  if (!pkgname) return { scope: null, name: null }
   const [_, scope = null, name = null] = (NAME_EXP.exec(pkgname) || [])
-  return {scope, name}
+  return { scope, name }
 }
 
