@@ -1,11 +1,9 @@
 'use strict'
 
-const docker = require('./docker/index.js')
+import * as docker from './docker/index.js'
 
-module.exports = postPublish
-
-async function postPublish(opts, context) {
-  const {logger} = context
+export async function postPublish(opts, context) {
+  const { logger } = context
   const image = docker.Image.from(opts, context)
   if (!opts.clean) return
 
