@@ -1,10 +1,8 @@
 'use strict'
 
-const docker = require('./docker/index.js')
+import * as docker from './docker/index.js'
 
-module.exports = publish
-
-async function publish(opts, context) {
+export async function publish(opts, context) {
   const image = docker.Image.from(opts, context)
   await image.push()
 }
