@@ -1,12 +1,10 @@
 'use strict'
 
-const path = require('path')
-const {promises: fs} = require('fs')
+import path from 'path'
+import fs from 'node:fs/promises'
 
-module.exports = getPkg
-
-async function getPkg(opts) {
-  const {cwd = process.cwd()} = opts || {}
+export async function readPkg(opts) {
+  const { cwd = process.cwd() } = opts || {}
   const pkg = await fs.readFile(path.join(cwd, 'package.json'), 'utf8')
   return JSON.parse(pkg)
 }
