@@ -1,25 +1,25 @@
 'use strict'
 
-const {promises: fs} = require('fs')
-const execa = require('execa')
-const debug = require('debug')('semantic-release:semantic-release-docker:verify')
-const SemanticError = require('@semantic-release/error')
-const prepare = require('./prepare.js')
-const docker = require('./docker/index.js')
+import fs from 'node:fs/promises'
+import { execa } from 'execa'
+import createDebug from 'debug'
+import SemanticError from '@semantic-release/error'
+import { dockerPrepare as prepare } from './prepare.js'
+import * as docker from './docker/index.js'
 
-module.exports = verify
+const debug = createDebug('semantic-release:semantic-release-docker:verify')
 
-async function verify(opts, context) {
-  const {env} = context
+export async function verify(opts, context) {
+  const { env } = context
   const PASSWORD = env.DOCKER_REGISTRY_PASSWORD || env.GITHUB_TOKEN
   const USERNAME = env.DOCKER_REGISTRY_USER
 
   if (!opts.name) {
     const error = new SemanticError(
       'Docker image name not found'
-    , 'EINVAL'
-    , 'Image name parsed from package.json name if possible. '
-        + 'Or via the "dockerImage" option.'
+      , 'EINVAL'
+      , 'Image name parsed from package.json name if possible. '
+      + 'Or via the "dockerImage" option.'
     )
     throw error
   }
@@ -33,9 +33,9 @@ async function verify(opts, context) {
   } catch (err) {
     const error = new SemanticError(
       `Unable to locate Dockerfile: ${image.dockerfile}`
-    , err.code
-    , 'Docker file is read from a local relative to PWD. Make sure the "dockerfile"'
-        + ' option is set to the desired location'
+      , err.code
+      , 'Docker file is read from a local relative to PWD. Make sure the "dockerfile"'
+      + ' option is set to the desired location'
     )
 
     throw error
@@ -45,7 +45,7 @@ async function verify(opts, context) {
   if (!USERNAME && !PASSWORD) {
     debug('No docker credentials found.  Skipping login')
   } else {
-    await doLogin({...opts, USERNAME, PASSWORD}, context)
+    await doLogin({ ...opts, USERNAME, PASSWORD }, context)
   }
 
   if (!opts.verifycmd) return true
@@ -58,8 +58,8 @@ async function verify(opts, context) {
 }
 
 async function doLogin(opts, context) {
-  const {USERNAME, PASSWORD} = opts
-  const {logger} = context
+  const { USERNAME, PASSWORD } = opts
+  const { logger } = context
 
   if (opts.login === false) {
     debug('docker login === false. Skipping login')
@@ -71,8 +71,8 @@ async function doLogin(opts, context) {
     if (set !== 2) {
       const error = new SemanticError(
         `Docker authentication failed ${USERNAME} ${PASSWORD}`
-      , 'EAUTH'
-      , 'Both ENV vars DOCKER_REGISTRY_USER and DOCKER_REGISTRY_PASSWORD must be set'
+        , 'EAUTH'
+        , 'Both ENV vars DOCKER_REGISTRY_USER and DOCKER_REGISTRY_PASSWORD must be set'
       )
       throw error
     }
@@ -80,9 +80,9 @@ async function doLogin(opts, context) {
     const passwd = execa('echo', [PASSWORD])
     const login = execa('docker', [
       'login'
-    , opts.registry || ''
-    , '-u', USERNAME
-    , '--password-stdin'
+      , opts.registry || ''
+      , '-u', USERNAME
+      , '--password-stdin'
     ])
 
     passwd.stdout.pipe(login.stdin)
@@ -93,8 +93,8 @@ async function doLogin(opts, context) {
       logger.fatal(err)
       const error = new SemanticError(
         `Docker authentication failed ${USERNAME} ${PASSWORD}`
-      , 'EAUTH'
-      , `Authentication to ${opts.registry || 'dockerhub'} failed`
+        , 'EAUTH'
+        , `Authentication to ${opts.registry || 'dockerhub'} failed`
       )
       throw error
     }
