@@ -1,9 +1,7 @@
 'use strict'
 
-const execa = require('execa')
+import { execa } from 'execa'
 
-module.exports = add
-
-async function add(cwd, file = '.') {
-  await execa('git', ['add', file], {cwd: cwd})
+export async function add(cwd, file = '.') {
+  await execa('git', ['add', file], { cwd: cwd })
 }
