@@ -1,11 +1,9 @@
 'use strict'
 
-const execa = require('execa')
-const head = require('./head.js')
+import { execa } from 'execa'
+import { head } from './head.js'
 
-module.exports = commit
-
-async function commit(cwd, message) {
-  await execa('git', ['commit', '-m', message], {cwd: cwd})
+export async function commit(cwd, message) {
+  await execa('git', ['commit', '-m', message], { cwd: cwd })
   return head(cwd)
 }
