@@ -1,10 +1,8 @@
 'use strict'
 
-const execa = require('execa')
+import { execa } from 'execa'
 
-module.exports = head
-
-async function head(cwd) {
-  const {stdout} = await execa('git', ['rev-parse', 'HEAD'], {cwd: cwd})
+export async function head(cwd) {
+  const { stdout } = await execa('git', ['rev-parse', 'HEAD'], { cwd: cwd })
   return stdout
 }
