@@ -1,13 +1,11 @@
 'use strict'
 
-const execa = require('execa')
-const initRemote = require('./init-remote.js')
+import { execa } from 'execa'
+import { initRemote } from './init-remote.js'
 
-module.exports = initOrigin
-
-async function initOrigin(cwd) {
+export async function initOrigin(cwd) {
   const origin = await initRemote()
-  await execa('git', ['remote', 'add', 'origin', origin], {cwd: cwd})
-  await execa('git', ['push', '--all', 'origin'], {cwd: cwd})
+  await execa('git', ['remote', 'add', 'origin', origin], { cwd: cwd })
+  await execa('git', ['push', '--all', 'origin'], { cwd: cwd })
   return origin
 }
