@@ -1,16 +1,14 @@
 'use strict'
 
-const path = require('path')
-const os = require('os')
-const {promises: fs} = require('fs')
-const execa = require('execa')
+import path from 'path'
+import os from 'os'
+import fs from 'node:fs/promises'
+import { execa } from 'execa'
 
-module.exports = initRemote
-
-async function initRemote(branch = 'main') {
+export async function initRemote(branch = 'main') {
   const cwd = await fs.mkdtemp(path.join(os.tmpdir(), path.sep))
   await execa('git', [
     'init', '--bare', `--initial-branch=${branch}`
-  ], {cwd: cwd})
+  ], { cwd: cwd })
   return `file://${cwd}`
 }
