@@ -1,13 +1,11 @@
 'use strict'
 
-const execa = require('execa')
+import { execa } from 'execa'
 
-module.exports = tag
-
-async function tag(cwd, name, hash) {
+export async function tag(cwd, name, hash) {
   const args = hash
     ? ['tag', '-f', name, hash]
     : ['tag', name]
 
-  await execa('git', args, {cwd: cwd})
+  await execa('git', args, { cwd: cwd })
 }
