@@ -1,14 +1,12 @@
 'use strict'
 
-const os = require('os')
-const execa = require('execa')
+import os from 'os'
+import { execa } from 'execa'
 
-module.exports = tags
-
-async function tags(cwd, hash) {
+export async function tags(cwd, hash) {
   const cmd = hash
     ? ['describe', '--tags', '--exact-match', hash]
     : ['tag', '-l', '--sort', 'v:refname']
-  const {stdout} = await execa('git', cmd, {cwd: cwd})
+  const { stdout } = await execa('git', cmd, { cwd: cwd })
   return stdout.split(os.EOL).filter(Boolean)
 }
