@@ -1,20 +1,20 @@
 'use strict'
 
-const path = require('path')
-const {test, threw} = require('tap')
-const buildConfig = require('../../lib/build-config.js')
+import path from 'path'
+import { test } from 'tap'
+import { buildConfig } from '../../lib/build-config.js'
 
 test('build-config', async (t) => {
   t.testdir({
     standard: {
-      'package.json': JSON.stringify({name: 'this-is-not-scoped'})
+      'package.json': JSON.stringify({ name: 'this-is-not-scoped' })
     }
-  , scoped: {
-      'package.json': JSON.stringify({name: '@scope/this-is-scoped'})
+    , scoped: {
+      'package.json': JSON.stringify({ name: '@scope/this-is-scoped' })
     }
-  , workspace: {
+    , workspace: {
       one: {
-        'package.json': JSON.stringify({name: '@internal/package'})
+        'package.json': JSON.stringify({ name: '@internal/package' })
       }
     }
   })
@@ -26,26 +26,26 @@ test('build-config', async (t) => {
     })
     tt.match(config, {
       dockerfile: 'Dockerfile'
-    , publish: true
-    , nocache: false
-    , tags: ['latest', '{{major}}-latest', '{{version}}']
-    , args: {
+      , publish: true
+      , nocache: false
+      , tags: ['latest', '{{major}}-latest', '{{version}}']
+      , args: {
         SRC_DIRECTORY: 'standard'
-      , TARGET_PATH: '.'
-      , NPM_PACKAGE_NAME: 'this-is-not-scoped'
-      , NPM_PACKAGE_SCOPE: null
-      , CONFIG_NAME: 'this-is-not-scoped'
-      , CONFIG_PROJECT: null
+        , TARGET_PATH: '.'
+        , NPM_PACKAGE_NAME: 'this-is-not-scoped'
+        , NPM_PACKAGE_SCOPE: null
+        , CONFIG_NAME: 'this-is-not-scoped'
+        , CONFIG_PROJECT: null
       }
-    , pkg: Object
-    , registry: null
-    , name: 'this-is-not-scoped'
-    , project: null
-    , build: 'id'
-    , context: '.'
-    , quiet: true
-    , clean: true
-    , dry_run: false
+      , pkg: Object
+      , registry: null
+      , name: 'this-is-not-scoped'
+      , project: null
+      , build: 'id'
+      , context: '.'
+      , quiet: true
+      , clean: true
+      , dry_run: false
     })
   })
 
@@ -56,31 +56,31 @@ test('build-config', async (t) => {
       options: {
         root: t.testdirName
       }
-    , cwd: path.join(t.testdirName, 'workspace', 'one')
+      , cwd: path.join(t.testdirName, 'workspace', 'one')
     })
     tt.match(config, {
       dockerfile: 'Dockerfile'
-    , nocache: false
-    , publish: true
-    , tags: ['latest', '{{major}}-latest', '{{version}}']
-    , platform: []
-    , args: {
+      , nocache: false
+      , publish: true
+      , tags: ['latest', '{{major}}-latest', '{{version}}']
+      , platform: []
+      , args: {
         SRC_DIRECTORY: 'one'
-      , TARGET_PATH: 'workspace/one'
-      , NPM_PACKAGE_NAME: 'package'
-      , NPM_PACKAGE_SCOPE: 'internal'
-      , CONFIG_NAME: 'package'
-      , CONFIG_PROJECT: 'internal'
+        , TARGET_PATH: 'workspace/one'
+        , NPM_PACKAGE_NAME: 'package'
+        , NPM_PACKAGE_SCOPE: 'internal'
+        , CONFIG_NAME: 'package'
+        , CONFIG_PROJECT: 'internal'
       }
-    , pkg: Object
-    , registry: null
-    , name: 'package'
-    , project: 'internal'
-    , build: 'id'
-    , context: '.'
-    , quiet: true
-    , dry_run: true
-    , clean: true
+      , pkg: Object
+      , registry: null
+      , name: 'package'
+      , project: 'internal'
+      , build: 'id'
+      , context: '.'
+      , quiet: true
+      , dry_run: true
+      , clean: true
     })
   })
 
@@ -92,98 +92,98 @@ test('build-config', async (t) => {
       })
       tt.match(config, {
         dockerfile: 'Dockerfile'
-      , nocache: false
-      , platform: []
-      , tags: ['latest', '{{major}}-latest', '{version}']
-      , args: {
+        , nocache: false
+        , platform: []
+        , tags: ['latest', '{{major}}-latest', '{version}']
+        , args: {
           SRC_DIRECTORY: 'scoped'
-        , TARGET_PATH: '.'
-        , NPM_PACKAGE_NAME: '@scope/this-is-scoped'
-        , NPM_PACKAGE_SCOPE: 'scope'
-        , CONFIG_NAME: 'this-is-scoped'
-        , CONFIG_PROJECT: 'scope'
+          , TARGET_PATH: '.'
+          , NPM_PACKAGE_NAME: '@scope/this-is-scoped'
+          , NPM_PACKAGE_SCOPE: 'scope'
+          , CONFIG_NAME: 'this-is-scoped'
+          , CONFIG_PROJECT: 'scope'
         }
-      , pkg: Object
-      , registry: null
-      , name: 'this-is-scoped'
-      , project: 'scope'
-      , build: 'id'
-      , context: '.'
-      , clean: true
-      , quiet: true
+        , pkg: Object
+        , registry: null
+        , name: 'this-is-scoped'
+        , project: 'scope'
+        , build: 'id'
+        , context: '.'
+        , clean: true
+        , quiet: true
       })
     }
 
     {
       const config = await buildConfig('id', {
         dockerProject: 'kittens'
-      , dockerImage: 'override'
-      , dockerFile: 'Dockerfile.test'
-      , dockerPublish: false
-      , dockerPlatform: 'linux/amd64'
-      , dockerBuildQuiet: 'false'
+        , dockerImage: 'override'
+        , dockerFile: 'Dockerfile.test'
+        , dockerPublish: false
+        , dockerPlatform: 'linux/amd64'
+        , dockerBuildQuiet: 'false'
       }, {
         cwd: path.join(t.testdirName, 'scoped')
       })
       tt.match(config, {
         dockerfile: 'Dockerfile.test'
-      , publish: false
-      , nocache: false
-      , platform: ['linux/amd64']
-      , tags: ['latest', '{{major}}-latest', '{{version}}']
-      , args: {
+        , publish: false
+        , nocache: false
+        , platform: ['linux/amd64']
+        , tags: ['latest', '{{major}}-latest', '{{version}}']
+        , args: {
           SRC_DIRECTORY: 'scoped'
-        , TARGET_PATH: '.'
-        , NPM_PACKAGE_NAME: '@scope/this-is-scoped'
-        , NPM_PACKAGE_SCOPE: 'scope'
-        , CONFIG_NAME: 'override'
-        , CONFIG_PROJECT: 'kittens'
+          , TARGET_PATH: '.'
+          , NPM_PACKAGE_NAME: '@scope/this-is-scoped'
+          , NPM_PACKAGE_SCOPE: 'scope'
+          , CONFIG_NAME: 'override'
+          , CONFIG_PROJECT: 'kittens'
         }
-      , pkg: Object
-      , registry: null
-      , name: 'override'
-      , project: 'kittens'
-      , build: 'id'
-      , context: '.'
-      , clean: true
-      , quiet: false
+        , pkg: Object
+        , registry: null
+        , name: 'override'
+        , project: 'kittens'
+        , build: 'id'
+        , context: '.'
+        , clean: true
+        , quiet: false
       })
     }
 
     {
       const config = await buildConfig('id', {
         dockerProject: null
-      , dockerImage: 'override'
-      , dockerFile: 'Dockerfile.test'
-      , dockerTags: 'latest,{{major}}-latest , fake, {{version}}'
-      , dockerAutoClean: false
-      , dockerPlatform: ['linux/amd64', 'linux/arm64']
-      , dockerBuildQuiet: 'false'
+        , dockerImage: 'override'
+        , dockerFile: 'Dockerfile.test'
+        , dockerTags: 'latest,{{major}}-latest , fake, {{version}}'
+        , dockerAutoClean: false
+        , dockerPlatform: ['linux/amd64', 'linux/arm64']
+        , dockerBuildQuiet: 'false'
       }, {
         cwd: path.join(t.testdirName, 'scoped')
       })
       tt.match(config, {
         dockerfile: 'Dockerfile.test'
-      , nocache: false
-      , platform: ['linux/amd64', 'linux/arm64']
-      , tags: ['latest', '{{major}}-latest', 'fake', '{{version}}']
-      , args: {
+        , nocache: false
+        , platform: ['linux/amd64', 'linux/arm64']
+        , tags: ['latest', '{{major}}-latest', 'fake', '{{version}}']
+        , args: {
           SRC_DIRECTORY: 'scoped'
-        , TARGET_PATH: '.'
-        , NPM_PACKAGE_NAME: '@scope/this-is-scoped'
-        , NPM_PACKAGE_SCOPE: 'scope'
-        , CONFIG_NAME: 'override'
-        , CONFIG_PROJECT: null
+          , TARGET_PATH: '.'
+          , NPM_PACKAGE_NAME: '@scope/this-is-scoped'
+          , NPM_PACKAGE_SCOPE: 'scope'
+          , CONFIG_NAME: 'override'
+          , CONFIG_PROJECT: null
         }
-      , pkg: Object
-      , registry: null
-      , name: 'override'
-      , project: null
-      , build: 'id'
-      , context: '.'
-      , quiet: false
-      , clean: false
+        , pkg: Object
+        , registry: null
+        , name: 'override'
+        , project: null
+        , build: 'id'
+        , context: '.'
+        , quiet: false
+        , clean: false
       })
     }
   })
-}).catch(threw)
+})
