@@ -1,8 +1,8 @@
 'use strict'
 
-const {test, threw} = require('tap')
-const buildConfig = require('../../lib/build-config.js')
-const buildTemplateVars = require('../../lib/build-template-vars.js')
+import { test } from 'tap'
+import { buildConfig } from '../../lib/build-config.js'
+import { buildTemplateVars } from '../../lib/build-template-vars.js'
 
 test('buildTemplateVars', async (t) => {
   const cwd = t.testdir({
@@ -12,53 +12,52 @@ test('buildTemplateVars', async (t) => {
   })
   const context = {
     cwd
-  , nextRelease: {
+    , nextRelease: {
       version: '1.0.0'
-    , gitTag: 'v1.0.0'
-    , gitHead: 'abcdefgh'
-    , type: 'major'
-    , notes: 'test it'
+      , gitTag: 'v1.0.0'
+      , gitHead: 'abcdefgh'
+      , type: 'major'
+      , notes: 'test it'
     }
   }
   const opts = await buildConfig('abacadaba', {
     dockerArgs: {
       TEMPLATE_VALUE: '{{type}}.{{version}}'
-    , BOOLEAN_VALUE: true
-    , NULL_VALUE: null
+      , BOOLEAN_VALUE: true
+      , NULL_VALUE: null
     }
   }, context)
 
   const vars = buildTemplateVars(opts, context)
   t.match(vars, {
     release_type: 'major'
-  , release_notes: 'test it'
-  , version: '1.0.0'
-  , git_sha: 'abcdefgh'
-  , git_tag: 'v1.0.0'
-  , pkg: {name: 'template-vars'}
-  , major: 1
-  , minor: 0
-  , patch: 0
-  , version: '1.0.0'
-  , network: 'default'
-  , next: {
-      major: 1
+    , release_notes: 'test it'
+    , version: '1.0.0'
+    , git_sha: 'abcdefgh'
+    , git_tag: 'v1.0.0'
+    , pkg: { name: 'template-vars' }
+    , major: 1
     , minor: 0
     , patch: 0
-    , version: '1.0.0'
+    , network: 'default'
+    , next: {
+      major: 1
+      , minor: 0
+      , patch: 0
+      , version: '1.0.0'
     }
-  , args: {
+    , args: {
       SRC_DIRECTORY: String
-    , TARGET_PATH: String
-    , NPM_PACKAGE_NAME: 'template-vars'
-    , NPM_PACKAGE_SCOPE: null
-    , CONFIG_NAME: String
-    , CONFIG_PROJECT: null
-    , GIT_SHA: 'abcdefgh'
-    , GIT_TAG: 'v1.0.0'
-    , TEMPLATE_VALUE: '{{type}}.{{version}}'
-    , BOOLEAN_VALUE: true
-    , NULL_VALUE: null
+      , TARGET_PATH: String
+      , NPM_PACKAGE_NAME: 'template-vars'
+      , NPM_PACKAGE_SCOPE: null
+      , CONFIG_NAME: String
+      , CONFIG_PROJECT: null
+      , GIT_SHA: 'abcdefgh'
+      , GIT_TAG: 'v1.0.0'
+      , TEMPLATE_VALUE: '{{type}}.{{version}}'
+      , BOOLEAN_VALUE: true
+      , NULL_VALUE: null
     }
   }, 'expected template values')
-}).catch(threw)
+})
