@@ -1,7 +1,7 @@
 'use strict'
 
-const {test, threw} = require('tap')
-const {gte} = require('../../../../lib/handlebars/helpers')
+import { test } from 'tap'
+import { gte } from '../../../../lib/handlebars/helpers/index.js'
 
 test('handlebars helpers', async (t) => {
   t.test('gte', async (t) => {
@@ -9,4 +9,4 @@ test('handlebars helpers', async (t) => {
     t.equal(gte(30, 40), false, '30 is not greater than or equal to 40)')
     t.equal(gte(50, 40), true, '40 is greater than or equal to 40)')
   })
-}).catch(threw)
+})
