@@ -1,7 +1,7 @@
 'use strict'
 
-const {test, threw} = require('tap')
-const {lte} = require('../../../../lib/handlebars/helpers')
+import { test } from 'tap'
+import { lte } from '../../../../lib/handlebars/helpers/index.js'
 
 test('handlebars helpers', async (t) => {
   t.test('gte', async (t) => {
@@ -9,4 +9,4 @@ test('handlebars helpers', async (t) => {
     t.equal(lte(40, 30), false, '40 is not less than or equal to 30')
     t.equal(lte(50, 50), false, '50 is less than or equal to 50)')
   })
-}).catch(threw)
+})
