@@ -1,27 +1,27 @@
 'use strict'
 
-const {test, threw} = require('tap')
-const parsePkgName = require('../../lib/parse-pkg-name.js')
+import { test } from 'tap'
+import { parsePkgName } from '../../lib/parse-pkg-name.js'
 
 test('parsePkgName', async (t) => {
-  t.deepEqual(parsePkgName('test'), {
+  t.same(parsePkgName('test'), {
     scope: null
-  , name: 'test'
+    , name: 'test'
   }, 'non-scoped package name')
 
-  t.deepEqual(parsePkgName('@namespace/foobar'), {
+  t.same(parsePkgName('@namespace/foobar'), {
     scope: 'namespace'
-  , name: 'foobar'
+    , name: 'foobar'
   }, 'scoped package name')
 
-  t.deepEqual(parsePkgName('nampace/foobar'), {
+  t.same(parsePkgName('nampace/foobar'), {
     name: null
-  , scope: null
+    , scope: null
   }, 'invalid package name')
 
-  t.deepEqual(parsePkgName(), {
+  t.same(parsePkgName(), {
     name: null
-  , scope: null
+    , scope: null
   }, 'no package name')
 
-}).catch(threw)
+})
