@@ -1,5 +1,3 @@
 'use strict'
 
-module.exports = {
-  Image: require('./image.js')
-}
+export { Image } from "./image.js";
