@@ -1,17 +1,16 @@
 'use strict'
 
-module.exports = {
-  endswith: require('./endswith')
-, eq: require('./eq')
-, gt: require('./gt')
-, gte: require('./gte')
-, includes: require('./includes')
-, lower: require('./lower')
-, lt: require('./lt')
-, lte: require('./lte')
-, neq: require('./neq')
-, pick: require('./pick')
-, startswith: require('./startswith')
-, split: require('./split')
-, upper: require('./upper')
-}
+export { endsWith as endswith } from './endswith.js'
+export { eq } from './eq.js'
+export { gt } from './gt.js'
+export { gte } from './gte.js'
+export { includes } from './includes.js'
+export { lower } from './lower.js'
+export { lt } from './lt.js'
+export { lte } from './lte.js'
+export { neq } from './neq.js'
+export { pick } from './pick.js'
+export { startsWith as startswith } from './startswith.js'
+export { split } from './split.js'
+export { upper } from './upper.js'
+
