@@ -1,5 +1,5 @@
 'use strict'
 
-module.exports = require('../../lang/array/to-array.js')
+export { toArray as split } from '../../lang/array/to-array.js'
 
 
