@@ -1,5 +1,3 @@
 'use strict'
 
-module.exports = {
-  toArray: require('./to-array.js')
-}
+export { toArray } from './to-array.js'
