@@ -1,6 +1,4 @@
 'use strict'
 
-module.exports = {
-  get: require('./get.js')
-, has: require('./has.js')
-}
+export { getProperty as get } from './get.js'
+export { hasProperty as has } from './has.js'
