@@ -1,6 +1,4 @@
 'use strict'
 
-module.exports = {
-  template: require('./template.js')
-, typecast: require('./typecast.js')
-}
+export { template } from './template.js'
+export { typeCast as typecast } from './typecast.js'
